Add ej:valid/2 for validation EJSON terms via declarative spec
Introduce the ability to validate EJSON objects using a declarative
spec shaped like EJSON, but with rules for validating instead of
literal values. The unit tests in test/ej_valid_test.erl show some
example usage.

This code working, but the API and format of the specs is subject to
change a bit.