Complex get/set now specified as {select, {Key, Value}}
The complex get/set behavior is different enough from the standard
mechanism that it is worth setting off with a tagged element in the
path. This makes the filtering explicit and will be more amenable to
extension in the future if desired.

It is now also possible to capture all objects in a list via {select,
all}. This fixes a problem with the top-level filtering which was too
aggressive and had introduced ambiguity where the same paths had
different meaning depending on the shape of the data.