Fix spec for nested ej:valid validations
I believe this last type was meant to capture the fact that ej:valid
supports nested specs like so:

    ej:valid({[
               {<<"name">>, string},
               {<<"modified_by">>, string},
               {<<"with">>, {[{<<"priority">>, number}]}}]}, JSON)

However, this call would be a violation of the spec as written.

Background:

Running dialyzer with Erlang 18.3 on chef-analytics (private github
repository) was yielding a `no local return` warning.  The cause
appeared to be a call to ej:valid that, when run in an erl shell, did in
fact return. Fixing this spec solves fixes the failure.

Signed-off-by: Steven Danna <steve@chef.io>